+++ src/lib/components/forms/checkbox-group/checkbox-group.component.spec.ts
@@ -5,7 +5,6 @@ import {
 } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { HttpClientModule } from '@angular/common/http';
-import { TooltipModule } from 'ngx-bootstrap';
 import { FormsModule } from '@angular/forms';
 import { LocalizationModule } from 'angular-l10n';
 import { l10nConfig } from '../../../../app/translation/l10n.config';
@@ -13,6 +12,7 @@ import { TerraMultiCheckBoxComponent } from '../multi-check-box/terra-multi-chec
 import { TerraMultiCheckBoxValueInterface } from '../multi-check-box/data/terra-multi-check-box-value.interface';
 import { TerraCheckboxComponent } from '../checkbox/terra-checkbox.component';
 import Spy = jasmine.Spy;
+import { TooltipDirective } from '../../tooltip/tooltip.directive';
 
 describe('Component: CheckboxGroupComponent', () =>
 {
@@ -39,12 +39,12 @@ describe('Component: CheckboxGroupComponent', () =>
     {
         TestBed.configureTestingModule(
             {
-                declarations: [CheckboxGroupComponent,
+                declarations: [TooltipDirective,
+                               CheckboxGroupComponent,
                                TerraMultiCheckBoxComponent,
                                TerraCheckboxComponent],
                 imports:      [
                     HttpClientModule,
-                    TooltipModule.forRoot(),
                     FormsModule,
                     LocalizationModule.forRoot(l10nConfig)
                 ]