+++ src/lib/components/info/terra-info.component.ts
@@ -11,6 +11,9 @@ import { TerraPlacementEnum } from '../../helpers/enums/terra-placement.enum';
 })
 export class TerraInfoComponent
 {
+    /**
+     * @deprecated since v4. Is replaced by the TooltipDirective and will be removed with the next major version.
+     */
     @Input()
     public textPlacement:TerraPlacementEnum;